@@ -1,4 +1,9 @@
-export const environment = {
-  production: true,
-  baseURL: 'http://localhost:9999/',
+import { Environment } from './environment.model';
+
+export const environment: Environment = {
+    production: true,
+    baseURL: 'http://localhost:9999/',
+    awsBucketName: `classifai-v2`,
+    awsAccessKeyId: `44LLO49S6VBXQ6APPW56`,
+    awsSecretAccessKey: `Diy1eCM9tNlzwqlwNS1eANSyucu2mePwJkdLXae0`,
 };
