@@ -3,13 +3,13 @@ import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
 const dataSetRoutes: Routes = [
-  {
-    path: '',
-    component: DataSetLayoutComponent,
-  },
+    {
+        path: '',
+        component: DataSetLayoutComponent,
+    },
 ];
 
 @NgModule({
-  imports: [RouterModule.forChild(dataSetRoutes)],
+    imports: [RouterModule.forChild(dataSetRoutes)],
 })
 export class DataSetRoutingModule {}
