@@ -1,31 +1,28 @@
-import { BehaviorSubject } from 'rxjs';
-import { Injectable } from '@angular/core';
+// import { BehaviorSubject } from 'rxjs';
+// import { Injectable } from '@angular/core';
 
-@Injectable({ providedIn: 'root' })
-export class ThemeService {
-  /** @state mainly used for set state */
-  private themeSubject = new BehaviorSubject<string>(null);
+// @Injectable({ providedIn: 'root' })
+// export class ThemeService {
+//     /** @state mainly used for set state */
+//     private themeSubject = new BehaviorSubject<string>('');
 
-  // /** @state mainly used for get state */
-  // theme$ = this.themeSubject.asObservable();
+//     // /** @state mainly used for get state */
+//     // theme$ = this.themeSubject.asObservable();
 
-  constructor() {}
+//     constructor() {}
 
-  setThemeState = (): void => {
-    const themeChecking = window.matchMedia('(prefers-color-scheme: light)')
-      .matches
-      ? 'dark-theme'
-      : 'light-theme';
-    // console.log(changedLangString);
-    this.themeSubject.next(themeChecking);
-    localStorage.setItem('theme', themeChecking);
-  };
+//     setThemeState = (): void => {
+//         const themeChecking = window.matchMedia('(prefers-color-scheme: light)').matches ? 'dark-theme' : 'light-theme';
+//         // console.log(changedLangString);
+//         this.themeSubject.next(themeChecking);
+//         localStorage.setItem('theme', themeChecking);
+//     };
 
-  getThemeState = (): string => {
-    const theme = localStorage.getItem('theme');
+//     getThemeState = (): string => {
+//         const theme = localStorage.getItem('theme');
 
-    /** @return {string} getValue() is used due to not an async request */
-    const themeResult = theme ? theme : this.themeSubject.getValue();
-    return themeResult ?? (this.setThemeState(), this.themeSubject.getValue());
-  };
-}
+//         /** @return {string} getValue() is used due to not an async request */
+//         const themeResult = theme ? theme : this.themeSubject.getValue();
+//         return themeResult ?? (this.setThemeState(), this.themeSubject.getValue());
+//     };
+// }
