@@ -1,63 +1,69 @@
 import { CacheBustingInterceptor } from './interceptors/cache-busting.interceptor';
-import { ClassifaiModalModule } from './classifai-modal/classifai-modal.module';
 import { CommonModule } from '@angular/common';
 import { FileNamePipe } from './pipe/file-name.pipe';
 import { FormatLanguagePipe } from './pipe/format-language.pipe';
-import { FormService } from './services/form.service';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { HTTP_INTERCEPTORS, HttpClient } from '@angular/common/http';
 import { LanguageService } from './services/language.service';
-import { LazyImgDirective } from './directives/lazy-image.directive';
+import { LazyLoadImgDirective } from './directives/lazy-image.directive';
+import { LoadingSpinnerInterceptor } from './interceptors/loading-spinner.interceptor';
+import { ModalModule } from 'src/components/modal/modal.module';
 import { NgModule } from '@angular/core';
-import { RouterModule } from '@angular/router';
-import { ThemeService } from './services/theme.service';
+import { PageHeaderModule } from 'src/components/page-header/page-header.module';
+import { SpinnerModule } from 'src/components/spinner/spinner.module';
 import { TranslateHttpLoader } from '@ngx-translate/http-loader';
 import { TranslateLoader, TranslateModule } from '@ngx-translate/core';
 
 // AoT requires an exported function for factories
-export function HttpLoaderFactory(httpClient: HttpClient) {
-  return new TranslateHttpLoader(httpClient, `../assets/i18n/`, `.json`);
-}
-
+export const httpLoaderFactory = (httpClient: HttpClient) => {
+    return new TranslateHttpLoader(httpClient, `../assets/i18n/`, `.json`);
+};
 @NgModule({
-  declarations: [FormatLanguagePipe, FileNamePipe, LazyImgDirective],
-  imports: [
-    RouterModule,
-    CommonModule,
-    FormsModule,
-    ReactiveFormsModule,
-    ClassifaiModalModule,
-    TranslateModule.forRoot({
-      loader: {
-        provide: TranslateLoader,
-        useFactory: HttpLoaderFactory,
-        deps: [HttpClient],
-      },
-    }),
-  ],
-  exports: [
-    FormatLanguagePipe,
-    FileNamePipe,
-    LazyImgDirective,
-    ClassifaiModalModule,
-    FormsModule,
-    ReactiveFormsModule,
-    TranslateModule,
-  ],
-  providers: [
-    LanguageService,
-    ThemeService,
-    FormService,
-    {
-      provide: HTTP_INTERCEPTORS,
-      useClass: CacheBustingInterceptor,
-      multi: true,
-    },
-    // {
-    //   provide: HTTP_INTERCEPTORS,
-    //   useClass: ResponseJsonInterceptor,
-    //   multi: true,
-    // },
-  ],
+    declarations: [
+        FormatLanguagePipe,
+        FileNamePipe,
+        // SanitizeSvgPipe,
+        LazyLoadImgDirective,
+    ],
+    imports: [
+        CommonModule,
+        FormsModule,
+        ReactiveFormsModule,
+        ModalModule,
+        SpinnerModule,
+        PageHeaderModule,
+        TranslateModule.forRoot({
+            loader: {
+                provide: TranslateLoader,
+                useFactory: httpLoaderFactory,
+                deps: [HttpClient],
+            },
+        }),
+    ],
+    exports: [
+        FormatLanguagePipe,
+        FileNamePipe,
+        // SanitizeSvgPipe,
+        ModalModule,
+        SpinnerModule,
+        PageHeaderModule,
+        FormsModule,
+        ReactiveFormsModule,
+        TranslateModule,
+        LazyLoadImgDirective,
+    ],
+    providers: [
+        LanguageService,
+        {
+            provide: HTTP_INTERCEPTORS,
+            useClass: CacheBustingInterceptor,
+            multi: true,
+        },
+        {
+            provide: HTTP_INTERCEPTORS,
+            useClass: LoadingSpinnerInterceptor,
+            multi: true,
+        },
+    ],
 })
 export class SharedModule {}
